Add NwArea.Deserialize. Fix crash from directory cleanup in ResourceManager.

+++ src/main/NWN/Services/ResourceManager/ResourceManager.cs
@@ -1,5 +1,8 @@
 using System;
 using System.IO;
+using System.Linq;
+using Anvil.Internal;
+using NLog;
 using NWN.API;
 using NWN.Native.API;
 using NWN.Plugins;
@@ -9,18 +12,28 @@ namespace NWN.Services
   [ServiceBinding(typeof(ResourceManager))]
   public sealed class ResourceManager : IDisposable
   {
+    public const int MaxNameLength = 16;
+
+    private static readonly Logger Log = LogManager.GetCurrentClassLogger();
+
     private const string AliasBaseName = "NMAN";
-    private const uint BasePriority = 70000000;
+    private const uint BasePriority = 70500000;
 
     private static readonly CExoBase ExoBase = NWNXLib.ExoBase();
     private static readonly CExoResMan ResMan = NWNXLib.ExoResMan();
+
+    private readonly CExoString TempAlias;
 
     private uint currentIndex;
 
     public ResourceManager(ITypeLoader typeLoader)
     {
-      Directory.Delete(EnvironmentConfig.ResourcePath, true);
-      CreateResourceDirectory(EnvironmentConfig.ResourcePath);
+      if (Directory.Exists(EnvironmentConfig.ResourcePath))
+      {
+        Directory.Delete(EnvironmentConfig.ResourcePath, true);
+      }
+
+      TempAlias = CreateResourceDirectory(EnvironmentConfig.ResourcePath).ToExoString();
 
       foreach (string resourcePath in typeLoader.ResourcePaths)
       {
@@ -28,20 +41,40 @@ namespace NWN.Services
       }
     }
 
-    private void CreateResourceDirectory(string path)
+    public void WriteTempResource(string resourceName, byte[] data)
+    {
+      string nameWithoutExtension = Path.GetFileNameWithoutExtension(resourceName);
+
+      if (nameWithoutExtension.Length > MaxNameLength)
+      {
+        throw new ArgumentOutOfRangeException(nameof(resourceName), $"Resource name (excl. extension) must be less than {MaxNameLength} characters.");
+      }
+
+      if (nameWithoutExtension.Any(c => !char.IsLetterOrDigit(c) && c != '_'))
+      {
+        throw new ArgumentOutOfRangeException(nameof(resourceName), $"Resource name must only contain alphanumeric characters, or underscores.");
+      }
+
+      File.WriteAllBytes(Path.Combine(EnvironmentConfig.ResourcePath, resourceName), data);
+    }
+
+    private string CreateResourceDirectory(string path)
     {
       if (string.IsNullOrEmpty(path))
       {
         throw new ArgumentOutOfRangeException(nameof(path), "Path must not be empty or null.");
       }
 
-      CExoString aliasName = (AliasBaseName + currentIndex).ToExoString();
+      string alias = AliasBaseName + currentIndex;
+      CExoString exoAlias = alias.ToExoString();
 
-      ExoBase.m_pcExoAliasList.Add(aliasName, path.ToExoString());
-      ResMan.CreateDirectory(aliasName);
-      ResMan.AddResourceDirectory(aliasName, BasePriority + currentIndex, true.ToInt());
+      Directory.CreateDirectory(path);
+      ExoBase.m_pcExoAliasList.Add(exoAlias, path.ToExoString());
+      ResMan.AddResourceDirectory(exoAlias, BasePriority + currentIndex, true.ToInt());
 
       currentIndex++;
+
+      return alias;
     }
 
     void IDisposable.Dispose()
